requirements.txt
wheel==0.45.1
aiohttp==3.11.13
aiosignal==1.3.1
async-timeout==4.0.2
asyncinit==0.2.4
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.115.12
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
multidict==6.0.4
openpyxl==3.0.10
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pydantic==1.10.4
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
python-on-whales==0.56.0
requests==2.28.2
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.46.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.12.2
urllib3==1.26.14
uvicorn==0.34.0
yarl==1.18.3
